Решение 2 урока

+++ src/lesson1/task1/Simple.kt
@@ -71,7 +71,7 @@ fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double =
  * Вывести значение того же угла в радианах (например, 0.63256).
  */
 fun angleInRadian(grad: Int, min: Int, sec: Int): Double =
-        (grad+(min+sec.toDouble()/60)/60.0)* PI/180
+        (grad+(min+sec/60.0)/60.0)* PI/180
 
 /**
  * Тривиальная
